create inscription components

+++ backend/app/app/api/api_v1/endpoints/student.py
@@ -45,19 +45,52 @@ def read_ancien_student(
     return all_student
 
 
-@router.get("/new/all", response_model=List[schemas.NewStudent])
+@router.get("/new/all", response_model=List[schemas.SelectStudentBase])
 def read_new_student(
         *,
         db: Session = Depends(deps.get_db),
         college_year: str,
+        uuid_mention: str,
         current_user: models.User = Depends(deps.get_current_active_user),
 ) -> Any:
     """
     Retrieve new student.
     """
-    students = crud.new_student.get_all(db=db, college_year=college_year)
-    return students
+    students = crud.new_student.get_by_mention(db=db,uuid_mention=uuid_mention, college_year=college_year)
+    all_student = []
+    for on_student in students:
+        on_student.mention = crud.mention.get_by_uuid(db=db, uuid=on_student.uuid_mention)
+        if find_in_list(current_user.uuid_mention, str(on_student.uuid_mention)) != -1:
+            all_student.append(on_student)
+    return all_student
 
+@router.get("/new_inscrit", response_model=List[schemas.NewStudent])
+def read_new_student(
+        *,
+        db: Session = Depends(deps.get_db),
+        college_year: str,
+        uuid_mention: str,
+        current_user: models.User = Depends(deps.get_current_active_user),
+) -> Any:
+    """
+    Retrieve new student.
+    """
+    students = crud.new_student.get_by_mention(db=db,uuid_mention=uuid_mention, college_year=college_year)
+    all_student = []
+    for on_student in students:
+        if on_student.uuid_journey:
+            for receipt in on_student.receipt_list:
+                receipt = receipt.replace("'",'"')
+                receipt = json.loads(receipt)
+                if receipt['year'] == on_student.actual_years:
+                    on_student.receipt = receipt
+                    break
+            stud = schemas.NewStudent(**jsonable_encoder(on_student))
+            stud.journey = crud.journey.get_by_uuid(db=db, uuid=on_student.uuid_journey)
+            on_student.mention = crud.mention.get_by_uuid(db=db, uuid=on_student.uuid_mention)
+            if find_in_list(current_user.uuid_mention, str(on_student.uuid_mention)) != -1:
+                all_student.append(on_student)
+    return all_student
 
 @router.post("/ancien", response_model=List[schemas.AncienStudent])
 def create_ancien_student(
@@ -116,11 +149,11 @@ def create_ancien_student(
     return all_student
 
 
-@router.post("/new", response_model=List[schemas.NewStudent])
+@router.post("/new", response_model=List[schemas.SelectStudentBase])
 def create_new_student(
         *,
         db: Session = Depends(deps.get_db),
-        student_in: schemas.SelectStudentBase,
+        student_in: schemas.SelectStudentCreate,
         current_user: models.User = Depends(deps.get_current_active_user),
 ) -> Any:
     """
@@ -131,21 +164,23 @@ def create_new_student(
     if not mention:
         raise HTTPException(status_code=404, detail="Mention not found")
 
-    student_in.num_select = f"S{mention.abbreviation.upper()}{student_in.num_select}{student_in.enter_year[2:4]}"
-    student_in.plugged = mention.plugged
     student = crud.new_student.get_by_num_select(db=db, num_select=student_in.num_select)
     if not student:
-        crud.new_student.create(db=db, obj_in=student_in)
+        student_in.num_select = f"S{mention.abbreviation.upper()}{student_in.num_select}-{student_in.enter_years[2:4]}"
+        student = crud.new_student.get_by_num_select(db=db, num_select=student_in.num_select)
+        if not student:
+            crud.new_student.create(db=db, obj_in=student_in)
+        else:
+            raise HTTPException(status_code=404, detail="Number already exist")
     else:
         crud.new_student.update(db=db, db_obj=student, obj_in=student_in)
     students = crud.new_student.get_all(db=db, college_year=student_in.actual_years)
     all_student = []
     for on_student in students:
-        if on_student.uuid_journey:
-            stud = schemas.NewStudent(**jsonable_encoder(on_student))
-            stud.journey = crud.journey.get_by_uuid(db=db, uuid=on_student.uuid_journey)
-            if find_in_list(current_user.uuid_mention, str(stud.uuid_mention)) != -1:
-                all_student.append(stud)
+        on_student.mention = crud.mention.get_by_uuid(db=db, uuid=on_student.uuid_mention)
+        if find_in_list(current_user.uuid_mention, str(on_student.uuid_mention)) != -1:
+            all_student.append(on_student)
+
     return all_student
 
 
@@ -154,6 +189,7 @@ def update_student_selected(
         *,
         db: Session = Depends(deps.get_db),
         student_in: schemas.NewStudentUpdate,
+        num_select: str,
         current_user: models.User = Depends(deps.get_current_active_user),
 ) -> Any:
     """
@@ -168,19 +204,38 @@ def update_student_selected(
     if not mention:
         raise HTTPException(status_code=404, detail="Mention not found")
 
-    student_in.plugged = mention.plugged
-    students = crud.new_student.get_student_admis(db=db, plugged=mention.plugged,
-                                                  college_year=student_in.actual_years)
+    students = crud.new_student.get_student_admis(db=db, college_year=student_in.actual_years,
+                                                  uuid_mention=mention.uuid)
     all_students = []
     for student in students:
         if student.num_carte:
             all_students.append(student)
     num = int(mention.last_num_carte) + len(all_students) + 1
 
-    student = crud.new_student.get_by_num_select(db=db, num_select=student_in.num_select)
+    student = crud.new_student.get_by_num_select(db=db, num_select=num_select)
+
+    if not student.receipt_list:
+        list_receipt = [str(dict(student_in.receipt))]
+        student_in.receipt_list = list_receipt
+    else:
+        k: int = 0
+        list_receipt = []
+        list_receipt_2 = []
+        for receipt in student.receipt_list:
+            receipt = receipt.replace("'",'"')
+            list_receipt.append(json.loads(receipt))
+        for receipt in list_receipt:
+            if receipt['year'] == student_in.receipt.year:
+                k = 1
+                list_receipt_2.append(str(dict(student_in.receipt)))
+            else:
+                list_receipt_2.append(str(dict(receipt)))
+        if k == 0:
+            list_receipt_2.append(str(dict(student_in.receipt)))
+        student_in.receipt_list = list_receipt_2
 
     if not student or not student.is_selected:
-        raise HTTPException(status_code=404, detail="Student not selected")
+        raise HTTPException(status_code=404, detail="Student not selected or not found")
 
     if student.num_carte:
         student_in.num_carte = student.num_carte
@@ -233,10 +288,10 @@ def read_student_by_num_select(
         current_user: models.User = Depends(deps.get_current_active_user),
 ) -> Any:
     """
-    Get student by num carte.
+    Get student by num select.
     """
     student = crud.new_student.get_by_num_select(db=db, num_select=num_select)
-    if not student:
+    if not student or not find_in_list(current_user.uuid_mention, str(student.uuid_mention)) != -1:
         raise HTTPException(status_code=404, detail="Student not found")
     return student
 
@@ -338,6 +393,7 @@ def delete_student(
         db: Session = Depends(deps.get_db),
         num_select: str,
         college_year: str,
+        uuid_mention: str,
         current_user: models.User = Depends(deps.get_current_active_user),
 ) -> Any:
     """
@@ -348,18 +404,14 @@ def delete_student(
         raise HTTPException(status_code=404, detail="Student not found")
 
     student = crud.new_student.remove_select(db=db, num_select=num_select)
-    if student:
-        print(os.path.exists(getcwd() + "/files/photos/" + student.photo))
+    if student and student.photo:
         if os.path.exists(getcwd() + "/files/photos/" + student.photo):
             os.remove(getcwd() + "/files/photos/" + student.photo)
-    students = crud.new_student.get_all(db=db, college_year=college_year)
+    students = crud.new_student.get_by_mention(db=db, uuid_mention=uuid_mention, college_year=college_year)
     all_student = []
     for on_student in students:
-        if on_student.uuid_journey:
-            stud = schemas.NewStudent(**jsonable_encoder(on_student))
-            stud.journey = crud.journey.get_by_uuid(db=db, uuid=on_student.uuid_journey)
-            if find_in_list(current_user.uuid_mention, str(stud.uuid_mention)) != -1:
-                all_student.append(stud)
+        if find_in_list(current_user.uuid_mention, str(on_student.uuid_mention)) != -1:
+            all_student.append(on_student)
     return all_student
